adjust 'getFailedPayments' lastPageFirstIndexOffset value
Adjust this value to be consistent with the value in the 'getPayments'
method since internally they both paginate in the exact same way. This
change makes the `next` value returned by this method consistent with
the 'getPayments' method.

The conditional check also checks for both '0' and '1' because the final
offset comes back as 1 if there are payments present and 0 if the lnd
has no payments at all in both methods.